fix Matrix pickle
